Add files via upload
The Objective of this evaluation is to appraise the implementation of software security in the scoped applications.  The Application Scope of this evaluation is WebGoat.  WebGoat is a web-based application developed in Java.  This test evaluated the applications for the following specific vulnerabilities related to Data Validation, Authentication & Authorization, Data Encryption, Error Handling & Logging, and Server Configuration: